collection layout  模式转换完成

+++ 仿百思不得姐OC/仿百思不得姐OC/Classes/Essence-精华/Controller/BSEssenceALLTableController.m
@@ -47,7 +47,7 @@
     {
         cell = [[UITableViewCell alloc] initWithStyle:UITableViewCellStyleSubtitle reuseIdentifier:ID];
     }
-    cell.textLabel.text = [NSString stringWithFormat:@"test-message-%ld", indexPath.row];
+    cell.textLabel.text = [NSString stringWithFormat:@"test-message-%ld", (long)indexPath.row];
     
     return cell;
 }
